requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
arq==0.26.1
redis==5.1.0
httpx==0.27.2
feedparser==6.0.11
google-generativeai==0.8.3
resend==2.4.0
structlog==24.4.0
jinja2==3.1.4
python-multipart==0.0.12
slowapi==0.1.9
python-dotenv==1.0.1